requirements.txt
python-dotenv
fastapi
requests
uvicorn
jinja2